Fix repo2docker version (#139)
It seems that a recent update to Mamba is causing a problem when building the
Docker image from the repo via repo2docker. See https://github.com/jupyterhub/mybinder.org-deploy/issues/2714.
This PR fixes the version of `repo2docker` to the last one that didn't have this problem,
and the image is built correctly.